requirements.txt
numpy==1.19.5
roboflow==1.1.14
pandas==1.3.3
duckietown-gym-daffy==6.1.34
duckietown-world-daffy==6.4.3
gym==0.17.1
Pillow=10.1.0